requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
django-nested-admin==3.3.2
djangorestframework==3.12.1
Pillow==7.2.0
pycodestyle==2.6.0
python-monkey-business==1.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
